refactor(viewer): extract useViewerExport from GeoViewer
Decomposition 4/4, content-neutral: capture registration, the per-slot
captureOnly re-render, the Safari-safe clipboard copy, export-dialog
state, and export-legend specs move into a hook. Chosen over the
originally planned chrome extraction — the sheet state is ~30 lines
inseparable from layout JSX a composition root legitimately owns.
GeoViewer docblock now describes the composition. Tests untouched
through all four extractions; GeoViewer 1590→905 lines.
